pflag: ignore invalid flags
In order to ensure that it is possible to use /etc/sysconfig/docker, we
have to ignore any undefined flags. Since the migrator doesn't require
any positional arguments, this means we can ignore all undefined
and left-over arguments.

We print warnings so the user is somewhat aware of what's going on.

Signed-off-by: Aleksa Sarai <asarai@suse.de>